tasklist tweeter

+++ app/Http/Controllers/TasklistsController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
-use App\Tasklist;    // add
+use App\Tasklist;
 
 class TasklistsController extends Controller
 {
@@ -15,11 +15,21 @@ class TasklistsController extends Controller
      */
     public function index()
     {
-        $tasklists = Tasklist::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasklists = $user->tasklists()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasklists.index', [
-            'tasklists' => $tasklists,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasklists' => $tasklists,
+            ];
+            $data += $this-> counts($user);
+            return view('users.show', $data);
+        }
+        else {
+            return view('welcome');
+        }
     }
 
 
@@ -46,13 +56,14 @@ class TasklistsController extends Controller
     public function store(Request $request)
     {
         $this->validate($request, [
-            'status' => 'required|max:10',   // added
+            'status' => 'required|max:10',   // add
             'content' => 'required|max:191',
         ]);
-        $tasklist = new Tasklist;
-        $tasklist->status = $request->status;    // added
-        $tasklist->content = $request->content;
-        $tasklist->save();
+
+        $request->user()->tasklists()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
 
         return redirect('/');
     }
@@ -114,9 +125,12 @@ class TasklistsController extends Controller
      */
     public function destroy($id)
     {
-        $tasklist = Tasklist::find($id);
-        $tasklist->delete();
+        $tasklist = \App\Tasklist::find($id);
 
-        return redirect('/');
+        if (\Auth::user()->id === $tasklist->user_id) {
+            $tasklist->delete();
+        }
+
+        return redirect()->back();
     }
 }
